Move auxiliary functions to their own files
