+++ src/Makefile
@@ -5,7 +5,7 @@
 
 CC ?= gcc
 CFLAGS += -Wall -w -O2 -D_FORTIFY_SOURCE=2 -fstack-protector --param=ssp-buffer-size=4 `pkg-config --cflags --libs gio-2.0 libffi` -g
-OBJ = dfuzzer.o introspection.o fuzz.o rand.o
+OBJ = dfuzzer.o introspection.o fuzz.o rand.o util.o
 TARGET = dfuzzer
 all: dfuzzer
 .PHONY: doc man clean install
